requirements.txt
torch
numpy
pandas
scikit-learn
matplotlib
PyYAML
wandb
pennylane
